abstract preloading strategy
