+++ biosimulations/libs/shared/utils/routes/src/index.ts
@@ -1,1 +1,4 @@
+import { MarkedPreloadingStrategy } from './lib/MarkedPreloadingStrategy';
+
 export * from './lib/routes.module';
+export { MarkedPreloadingStrategy } from './lib/MarkedPreloadingStrategy';
