+++ biosimulations/libs/shared/utils/routes/src/lib/routes.module.ts
@@ -1,7 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
+import { MarkedPreloadingStrategy } from './MarkedPreloadingStrategy';
+export const MARKED_PRELOADING_STRATEGY = MarkedPreloadingStrategy;
 @NgModule({
   imports: [CommonModule],
+  providers: [MARKED_PRELOADING_STRATEGY],
 })
 export class RoutesModule {}
